+++ src/main/java/novamachina/exnihiloae/common/init/ExNihiloAEItems.java
@@ -6,20 +6,20 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
-import novamachina.exnihiloae.common.utility.Constants;
+import novamachina.exnihiloae.common.utility.ExNihiloAEConstants;
 import novamachina.exnihilosequentia.common.init.ModInitialization;
 import novamachina.exnihilosequentia.common.utility.ExNihiloLogger;
 import org.apache.logging.log4j.LogManager;
 
-public class ModItems {
+public class ExNihiloAEItems {
     private static final ExNihiloLogger logger = new ExNihiloLogger(LogManager.getLogger());
     private static final DeferredRegister<Item> ITEMS = DeferredRegister.create(
-        ForgeRegistries.ITEMS, Constants.ModIds.EX_NIHILO_AE);
+        ForgeRegistries.ITEMS, ExNihiloAEConstants.ModIds.EX_NIHILO_AE);
     public static final RegistryObject<Item> CRUSHED_SKYSTONE = ITEMS
-        .register(Constants.Blocks.CRUSHED_SKYSTONE,
-            () -> new BlockItem(ModBlocks.CRUSHED_SKYSTONE.get(),
+        .register(ExNihiloAEConstants.Blocks.CRUSHED_SKYSTONE,
+            () -> new BlockItem(ExNihiloAEBlocks.CRUSHED_SKYSTONE.get(),
                 new Item.Properties().group(ModInitialization.ITEM_GROUP)));
-    private ModItems() {
+    private ExNihiloAEItems() {
     }
 
     public static void init(IEventBus modEventBus) {