+++ src/main/java/novamachina/exnihiloae/common/utility/ExNihiloAEConstants.java
@@ -1,6 +1,6 @@
 package novamachina.exnihiloae.common.utility;
 
-public class Constants {
+public class ExNihiloAEConstants {
     public class ModIds {
         public static final String EX_NIHILO_AE = "exnihiloae";