+++ changelog.md
@@ -1,1 +1,2 @@
 1.0.0.0
+- Add Crushed Sky Stone, Add recipes for AE2